+++ src/embeds/creators/skills.js
@@ -1,8 +1,8 @@
 // @flow
 
-import type { EmbedProps } from './bootstrap';
+import type { EmbedProps } from 'embeds/bootstrap';
 
-import Skills from './components/Skills';
+import Skills from 'embeds/components/Skills';
 
 export default function (element: HTMLElement, ids: Array<number>) {
   return (props: EmbedProps) => <Skills {...props} ids={ids} />;